Add files via upload
This was a quick 2D platformer game I made the week before Thanksgiving, basically it's a character with my Boss's face running around chopping turkeys with an ax. You can play it at Itch.io/kodierer the name is Turkey Hunter. Feel free to improve it, add extra levels and anything else you like. It's free to the world.